requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pytest
httpx
pydantic